Persist preference store to localstorage
