+++ apps/web/src/components/Common/Layout.tsx
@@ -15,9 +15,9 @@ import type { FC, ReactNode } from "react";
 import { useEffect } from "react";
 import { Toaster } from "react-hot-toast";
 import { useNonceStore } from "src/store/non-persisted/useNonceStore";
-import { usePreferencesStore } from "src/store/non-persisted/usePreferencesStore";
 import { useProfileStatus } from "src/store/non-persisted/useProfileStatus";
 import { hydrateAuthTokens, signOut } from "src/store/persisted/useAuthStore";
+import { usePreferencesStore } from "src/store/persisted/usePreferencesStore";
 import { useProfileStore } from "src/store/persisted/useProfileStore";
 import { useProfileThemeStore } from "src/store/persisted/useProfileThemeStore";
 import { isAddress } from "viem";
