+++ apps/web/src/components/Settings/Preferences/AppIcon.tsx
@@ -10,7 +10,7 @@ import axios from "axios";
 import type { FC } from "react";
 import { useState } from "react";
 import toast from "react-hot-toast";
-import { usePreferencesStore } from "src/store/non-persisted/usePreferencesStore";
+import { usePreferencesStore } from "src/store/persisted/usePreferencesStore";
 
 const icons = [
   { id: 0, name: "Default" },
