--- apps/web/src/store/non-persisted/usePreferencesStore.ts
@@ -1,50 +1,0 @@
-import { createTrackedSelector } from "react-tracked";
-import { create } from "zustand";
-
-interface State {
-  appIcon: number;
-  email: null | string;
-  emailVerified: boolean;
-  hasDismissedOrMintedMembershipNft: boolean;
-  highSignalNotificationFilter: boolean;
-  developerMode: boolean;
-  resetPreferences: () => void;
-  setAppIcon: (appIcon: number) => void;
-  setEmail: (email: null | string) => void;
-  setEmailVerified: (emailVerified: boolean) => void;
-  setHasDismissedOrMintedMembershipNft: (
-    hasDismissedOrMintedMembershipNft: boolean
-  ) => void;
-  setHighSignalNotificationFilter: (
-    highSignalNotificationFilter: boolean
-  ) => void;
-  setDeveloperMode: (developerMode: boolean) => void;
-  loading: boolean;
-  setLoading: (loading: boolean) => void;
-}
-
-const store = create<State>((set) => ({
-  appIcon: 0,
-  email: null,
-  emailVerified: false,
-  hasDismissedOrMintedMembershipNft: true,
-  highSignalNotificationFilter: false,
-  developerMode: false,
-  resetPreferences: () =>
-    set(() => ({
-      hasDismissedOrMintedMembershipNft: true,
-      highSignalNotificationFilter: false
-    })),
-  setAppIcon: (appIcon) => set(() => ({ appIcon })),
-  setEmail: (email) => set(() => ({ email })),
-  setEmailVerified: (emailVerified) => set(() => ({ emailVerified })),
-  setHasDismissedOrMintedMembershipNft: (hasDismissedOrMintedMembershipNft) =>
-    set(() => ({ hasDismissedOrMintedMembershipNft })),
-  setHighSignalNotificationFilter: (highSignalNotificationFilter) =>
-    set(() => ({ highSignalNotificationFilter })),
-  setDeveloperMode: (developerMode) => set(() => ({ developerMode })),
-  setLoading: (loading) => set(() => ({ loading })),
-  loading: false
-}));
-
-export const usePreferencesStore = createTrackedSelector(store);
